codex/requirements.txt
typer==0.12.3
pyyaml==6.0.2
requests==2.32.3
tenacity==9.0.0
rich==13.7.1
